requirements.txt
fastmcp
openai>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.10
uvicorn>=0.22.0
